+++ src/message.js
@@ -3,7 +3,7 @@
 import type { StatePath } from "./storage";
 
 /**
- * Tag identifying the message, used to subscribe and match messages.
+ * Tag identifying the message, used to match and refine messages.
  */
 export type MessageTag = string;
 
@@ -25,11 +25,6 @@ export type Message = {
 export type InflightMessage = {
   +_message: Message,
   +_source: StatePath,
-  /**
-   * If an active subscription has received this message this is the state path
-   * which received it.
-   */
-  _received: boolean,
 };
 
 /**