+++ test/storage.broadcast.test.js
@@ -2,8 +2,7 @@
 
 import ninos from "ninos";
 import ava from "ava";
-import { Storage } from "../src/storage";
-import { updateData, updateAndSend } from "../src/update";
+import { Storage, updateData, updateAndSend } from "../src";
 import { args, unhandledMessageError } from "./util";
 
 const test = ninos(ava).serial;
@@ -45,13 +44,11 @@ test("broadcastMessage triggers attempts to send messages to all states", t => {
   const emit = t.context.spy(s, "emit");
   const error = t.context.spy(console, "error", () => {});
   const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({}));
+  const update = t.context.stub(() => null);
   const d = {
     id: "foo",
     init,
     update,
-    subscribe,
   };
 
   const foo = s.createState(d);
@@ -65,11 +62,10 @@ test("broadcastMessage triggers attempts to send messages to all states", t => {
     ["messageQueued", { tag: "AAA" }, ["@"]],
     ["unhandledMessage", { tag: "AAA" }, ["@"]],
   ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [1],
+  t.deepEqual(args(update), [
+    [1, { tag: "AAA" }],
+    [1, { tag: "AAA" }],
   ]);
-  t.deepEqual(args(update), []);
   t.deepEqual(args(error), [
     unhandledMessageError({ tag: "AAA" }, ["@"]),
   ]);
@@ -80,13 +76,11 @@ test("broadcastMessage sends a message to all states with deepest first", t => {
   const emit = t.context.spy(s, "emit");
   const error = t.context.spy(console, "error", () => {});
   const init = t.context.stub(() => updateData(1));
-  const update = t.context.stub(() => updateData(2));
-  const subscribe = t.context.stub(() => ({ AAA: true }));
+  const update = t.context.stub((_, { tag }) => tag === "AAA" ? updateData(2) : null);
   const d = {
     id: "foo",
     init,
     update,
-    subscribe,
   };
 
   const foo = s.createState(d);
@@ -107,15 +101,11 @@ test("broadcastMessage sends a message to all states with deepest first", t => {
     ["messageMatched", { tag: "AAA" }, ["foo"]],
     ["stateNewData", 2, ["foo"], { tag: "AAA" }],
   ]);
-  t.deepEqual(args(subscribe), [
-    [1],
-    [2],
-    [1],
-    [2],
-    [1],
-    [2],
+  t.deepEqual(args(update), [
+    [1, { tag: "AAA" }],
+    [1, { tag: "AAA" }],
+    [1, { tag: "AAA" }],
   ]);
-  t.is(update.calls.length, 3);
   t.deepEqual(args(error), []);
 });
 
@@ -125,12 +115,10 @@ test("broadcastMessage propagates messages in order", t => {
   const error = t.context.spy(console, "error", () => {});
   const init = t.context.stub(() => updateData(1));
   const update = t.context.stub(() => updateAndSend(2, { tag: "BBB" }));
-  const subscribe = t.context.stub(() => ({ AAA: true, BBB: true }));
   const d = {
     id: "foo",
     init,
     update,
-    subscribe,
   };
 
   const foo = s.createState(d);